+++ iftttie/services/base.py
@@ -1,13 +1,18 @@
 from __future__ import annotations
 
 from abc import ABCMeta, abstractmethod
+from typing import AsyncIterable
+
+from iftttie.types_ import Event
 
 
 class Service(metaclass=ABCMeta):
     @property
     @abstractmethod
-    async def events(self):
+    async def events(self) -> AsyncIterable[Event]:
         """
         Gets an asynchronous generator of events.
         """
         raise NotImplementedError()
+        # noinspection PyUnreachableCode
+        yield NotImplemented